Make step-0.3 readable and positive (fixes #1755, fixes #1567)
Before the step was very imperative, to the point that it wasn't clear
whether or not it was a mandatory step. There were also typos and
redundancies.

The proposed changes shorten several sentences, make it clear that it is
an optional step, and makes the language more positive by replacing phrases
like *you may* with *it would help if you*...