Remove redundancy and improve structure of vi-faq Q8 (fixes 1564)
Before there were smatterings of sentences with essentially the same two
messages: (1) You're internship doesn't officially start until you're done
with the 10 steps, and (2) once you are done with the first steps, we'll
help you with your paperwork.

The proposed changes shorten the section, and split it into two logical
sections, the first of which deals with why ole doesn't sign paperwork
before the first steps are done and the second of which reassures that
one those steps are complete ole will lend its support to the vi.