Fix techgenius tables (fixes #1761)
Currently there's a redundant label and missing information in the techgenius
team page. We know that this is the techgenius page, so it doesn't need to be
listed in the table. Additionally the tech geniuses table at the bottom
of the page is empty, while the current members section is populated, which
makes no sense.

This commit removes the unecessary label and adds members to the tech genius
to the table.