Re-Organized Docker-Compose Section of Tutorial
There were several other issues related to ruining one's
planet configuration due to a variety of reasons.

> For example: *issue #2551*

I ran into the issue when using the docker-compose command
found in this section of the tutorial. The solution in
this commit is to simply move the warning text to be seen
before the volatile docker-compose command that can *break*
planet-configs.