Repositioned warning message in docker tutorial
There have been other issues related to destroying pre-
existing planet configs.

> For example, *see issue #2551*

I ran into this issue during the docker-compose section
of the docker tutorial. There is a warning message, but it
is positioned after a volatile docker-compose command
that **destroys** pre-existing planet configs.

This solution is simply a re-positioning of that warning to be
befor the volatile command is listed.